fix: rename enum/index.js -> lib/constant.js
