+++ packages/create-app/lib/constant.js
