+++ packages/create-app/lib/server.js
@@ -1,6 +1,6 @@
 import { $, which } from 'zx';
 import semver from 'semver';
-import { BLOCKLET_COMMAND } from '../enums/index.js';
+import { BLOCKLET_COMMAND } from './constant.js';
 
 $.verbose = false;
